Move Iceberg reader helpers to root-safe module
